Aparecendo botão de deletar só para aquele usuário.

+++ application/views/question/index.php
@@ -42,7 +42,9 @@ $name = $session['name'];
 					</div>
 					<div class="group-send-home" id="button-actions-home">
 						<a href="<?php echo base_url(); ?>index.php/response/index/<?php echo $question['id_question'] ?>/<?php echo $question['id_category'] ?>" style="border: 0; background: transparent;"><img src="<?php echo base_url(); ?>assets/icons/message.svg" /></a>
-						<a href="<?php echo base_url(); ?>index.php/question/destroy/<?php echo $question['id_question'] ?>/<?php echo $question['id_category'] ?>" style="border: 0; background: transparent;"><img src="<?php echo base_url(); ?>assets/icons/delete.svg" /></a>
+						<?php if ($session['id_user'] == $question['id_user']) { ?>
+							<a href="<?php echo base_url(); ?>index.php/question/destroy/<?php echo $question['id_question'] ?>/<?php echo $question['id_category'] ?>" style="border: 0; background: transparent;"><img src="<?php echo base_url(); ?>assets/icons/delete.svg" /></a>
+						<?php } ?>
 					</div>
 				</div>
 			</div>
